requirements.txt
ollama>=0.1.7
streamlit>=1.28.0
flask>=2.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
pdfminer.six>=20240506
scikit-learn>=1.7.0
scipy>=1.10.0
joblib>=1.3.0
